+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 0.1.0
 
-* Horizontal layout
+* Horizontal and vertical layout
 * Shrink weight
 * Shrink order
+* Main alignment
+* Cross alignment
 
 ## 0.0.1